distrib: introduce APKBUILD to make alpine linux packages
When added to the community aports[1], the APKBUILD file provides
a recipe to make a minimalistic, yet working package.
The "kb-docs" package will accompany the executable, and contain
only the license file (as MIT license requires).

apkbuild-lint is passed, the resulting package is installable
from a local repository, tested on alpine v3.21.

[1] https://gitlab.alpinelinux.org/alpine/aports#alpine-linux-aports-repository

Signed-off-by: xandrow indira <xandrow89@gmail.com>
Signed-off-by: ALex Nikonov <alex+redot@nikonov.tech>